requirements.txt
build~=1.3.0
coverage~=7.12.0
cupy-cuda13x~=13.6.0
flake8~=7.3.0
gpugt~=0.0.0.dev0
interrogate~=1.7.0
matplotlib~=3.10.6
noregret~=0.0.0.dev1
numpy~=2.3.5
open-spiel~=1.6.9
ordered-set~=4.1.0
pandas~=2.3.2
pokerkit~=0.7.0a4
scipy~=1.16.3
seaborn~=0.13.2
Sphinx~=8.2.3
sphinx-rtd-theme~=3.0.2
tqdm~=4.67.1
twine~=6.2.0
